delete xuat hang
