+++ lib/src/features/dashboard/screens/add/xuathang/widget/delete_item_xuathang.dart
@@ -1,24 +1,25 @@
 import 'package:flutter/material.dart';
-import 'package:hobin_warehouse/src/constants/color.dart';
+
 import '../../../../../../common_widgets/snackbar/snackbar.dart';
+import '../../../../../../constants/color.dart';
 import '../../../../../../constants/icon.dart';
 import '../../widget/card_add_widget.dart';
-import 'itemdachon_nhaphang.dart';
+import 'itemdachon_xuathang_widget.dart';
 
-class DeleteItemsScreen extends StatefulWidget {
+class DeleteItemXuatHangScreen extends StatefulWidget {
   final Map<String, dynamic> thongTinItemNhapHienTai;
   final Function onDelete;
-  const DeleteItemsScreen({
-    super.key,
-    required this.thongTinItemNhapHienTai,
-    required this.onDelete,
-  });
+  const DeleteItemXuatHangScreen(
+      {super.key,
+      required this.thongTinItemNhapHienTai,
+      required this.onDelete});
 
   @override
-  State<DeleteItemsScreen> createState() => _DeleteItemsScreenState();
+  State<DeleteItemXuatHangScreen> createState() =>
+      _DeleteItemXuatHangScreenState();
 }
 
-class _DeleteItemsScreenState extends State<DeleteItemsScreen> {
+class _DeleteItemXuatHangScreenState extends State<DeleteItemXuatHangScreen> {
   final List<Map<String, dynamic>> list = [];
   @override
   void initState() {
@@ -30,13 +31,13 @@ class _DeleteItemsScreenState extends State<DeleteItemsScreen> {
   Widget build(BuildContext context) {
     return Container(
       color: whiteColor,
-      height: MediaQuery.of(context).size.height * 0.30,
+      height: MediaQuery.of(context).size.height * 0.37,
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
           const SizedBox(height: 10),
-          ItemDaChonNhapHangWidget(
-            selectedItems: [widget.thongTinItemNhapHienTai],
+          ItemDaChonXuatHangWidget(
+            selectedItems: list,
             blockOnPress: true,
             reLoad: () {},
           ),
